requirements.txt
altgraph==0.17.2
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
Django==4.2.16
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.54.1
future==0.18.2
idna==3.10
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
macholib==1.15.2
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pdfkit==1.0.0
pillow==11.0.0
pipenv==2024.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.15.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
zipp==3.20.2
